Add home version one
I have been working on the homepage for the smallest height of mobile
screen 568px and then in the other mobile screens I will have media
queries because I just want to have the links to be displayed in the
screen height without scrolling down. The colours, font family and
sizes aren’t confirmed and also the header needs to be changed.