Add about page and text
I have created an about page and linked it with home. I have all the
pages working except work page which I haven’t created it. I haven’t
changed the font family yet and still need to work on the paragraphs.
Also the image there is going to be a portrait of me, maybe with a bit
different sizes (I think I want one with less height- around 3/4 of
this one) but I just decided to add this as a placeholder.