Continue work section
I have added the 2 projects but still need to link the random quote
machine with codepen. I want to remove the space between the first
project and 2nd project and to ask how the position of the footer keeps
changing and it is not sticked to the end of the page.