Add social media icons for contact page
I have done the svg icons and just need to center them for mobile
first. I added a width for the parent div that contains the social
icons and I set it to display flex but still not sure why it is not
centered yet. Will fix it next time working on the project. I have the
form to do next.