Add contact form version two
I have added the styling and how many rows and cols for the textarea
but then I decided to not have a contact form for now. I want to have a
really simple website to start applying for jobs and I already have my
email and my social networks’ accounts in the contact page so that is
will do the job for now.