Update projects images and add fonts
I updated the Random Quote Machine project image to the latest updates
I have done to it. Changed the size of the image of the second project
as well. I added 2 fonts, 1 for the body text and another for headlines
like the first line of about page and contact page get in touch
sentence. I added the links in the html and removed @import after
watching a tutorial about how this keyword makes things slower. I also
have my CV in a new folder named documents and I am still not sure how
I am going to add it to my portfolio. Still have a todo list of adds
and updates.