Add a new header style for big screens
For tablets and desktops I want to have a header that contains logo on
the left and navigation on the right. I used flexbox to do that. I will
just adjust the margin on the right and left of the header and the
padding between each nav-link for bigger screens like desktops and big
desktop screens.