Add rest of media queries 1024px for work and contact
I worked on the work page and displayed the projects inline-block and
set a padding between both. Also I adjusted the padding for the social
media to be centred a bit comparing to the contact-info section.  I now
have the big screens, 2 more media queries to go so my website would be
responsive as much as it could be.