Add rest of media queries version two
I done the work and contact pages. Still have few problems I need to
solve with the whole site and of course have my logo, change the image
in about section with my photo and finally add something to the
homepage. Not sure yet about adding my CV there or not.