Fix footer in small screens and tablets
I fixed the footer to be at the end of the page if the page has small
content, long content will be pushed at the end of the page. I also
wrapped the body of each page with a div that has class body and give
it padding bottom same number as the footer.